third_party/WebKit/LayoutTests/webaudio/AudioParam/value-setter-warnings-expected.txt

CONSOLE ERROR: line 39: [audit.js] this test requires the explicit comparison with the expected result when it runs with run_web_tests.py.
CONSOLE MESSAGE: line 129: Test: test-0
CONSOLE MESSAGE: line 129: Test: test-1
CONSOLE MESSAGE: line 129: Test: test-2
CONSOLE MESSAGE: line 129: Test: test-3
CONSOLE MESSAGE: line 129: Test: test-4
CONSOLE MESSAGE: line 129: Test: test-5
This is a testharness.js-based test.
PASS # AUDIT TASK RUNNER STARTED.
PASS > [test-0] 
PASS   test-0 : rendered successfully
PASS < [test-0] All assertions passed. (total 1 assertions)
PASS > [test-1] 
PASS   test-1 : rendered successfully
PASS < [test-1] All assertions passed. (total 1 assertions)
PASS > [test-2] 
PASS   test-2 : rendered successfully
PASS < [test-2] All assertions passed. (total 1 assertions)
PASS > [test-3] 
PASS   test-3 : rendered successfully
PASS < [test-3] All assertions passed. (total 1 assertions)
PASS > [test-4] 
PASS   test-4 : rendered successfully
PASS < [test-4] All assertions passed. (total 1 assertions)
PASS > [test-5] 
PASS   Value setter overlaps setValueCurve threw NotSupportedError: "Failed to set the 'value' property on 'AudioParam': setValueAtTime(99, 0.4986666666666666) overlaps setValueCurveAtTime(..., 0.25, 0.5)".
PASS   test-5 : rendered successfully
PASS < [test-5] All assertions passed. (total 2 assertions)
PASS # AUDIT TASK RUNNER FINISHED: 6 tasks ran successfully.
Harness: the test ran to completion.
